Refactor: API 리팩토링
Refactor: API 리팩토링

API, Response, Request 파일 세분화

이슈 유형: #64
