+++ Duckmelang/Duckmelang/MyAccompany/ViewControllers/MyPostsViewController.swift
@@ -9,7 +9,7 @@ import UIKit
 import Moya
 
 class MyPostsViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
-    private let provider = MoyaProvider<AllEndpoint>(plugins: [NetworkLoggerPlugin(configuration: .init(logOptions: .verbose))])
+    private let provider = MoyaProvider<MyAccompanyAPI>(plugins: [NetworkLoggerPlugin(configuration: .init(logOptions: .verbose))])
     private var myPostsData: [PostDTO] = []
     
     override func viewDidLoad() {
